backend/requirements.txt
asgiref==3.7.2
attrs==22.1.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.5
crispy-bootstrap5==2023.10
cycler==0.11.0
distlib==0.3.8
Django==5.0.2
django-cors-headers==3.13.0
django-crispy-forms==2.1
djangorestframework==3.14.0
exceptiongroup==1.0.4
filelock==3.13.1
fonttools==4.37.4
idna==3.4
image==1.5.33
iniconfig==1.1.1
kiwisolver==1.4.4
matplotlib==3.6.1
nose==1.3.7
numpy==1.23.4
oauthlib==3.2.2
packaging==21.3
Pillow==9.2.0
pipenv==2023.12.1
platformdirs==4.2.0
pluggy==1.0.0
pyparsing==3.0.9
pytest==7.2.0
python-dateutil==2.8.2
pytz==2024.1
qrcode==7.3.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sortedcontainers==2.4.0
sqlparse==0.4.3
tomli==2.0.1
tweepy==4.12.1
typing_extensions==4.9.0
tzdata==2022.5
urllib3==1.26.13
urwid==2.1.2
virtualenv==20.25.1
virtualenvwrapper-win==1.2.7
